Added the About page content.
So far so good, added the content page as I was following along. Looking pretty nice with some simple CSS and content. I think I am starting to see how I can make this better with SASS and better consistency in the CSS framework, so I cannot wait to get there in the learning!